Rename a utility method
And add a few more assertions in test
