+++ src/utils/index.js
@@ -2,12 +2,12 @@ import * as mixins from './mixins';
 import vuetifyConfig from './vuetifyConfig';
 import Upload from './upload';
 import DebounceCounter from './debouncer';
-import getLocationValidator from './getLocationValidator';
+import createLocationValidator from './createLocationValidator';
 
 export {
   mixins,
   DebounceCounter,
   vuetifyConfig,
   Upload,
-  getLocationValidator,
+  createLocationValidator,
 };
